Submit test case for squigly heredoc bug
I'm noticing rufo butcher public keys. I was only planning on adding the
comment test, but I'm unable to get any of the three tests to pass at
all.

The test complains the results are indented, so I outdent the expected,
and then it complains they're outdented. It seems impossible to satisfy.

I don't have any strong feelings about how much indent is right, only
that it should be the same for all three examples and that the public
key shouldn't have it's first line indented beyond the rest of the
document.

This looks like it might be overlapping with https://github.com/ruby-formatter/rufo/issues/93